# crypto/internal/fips140/edwards25519/field: speed up Element.Bytes